requirements.txt
streamlit
langchain
pypdf2
python-dotenv
faiss-cpu